spark/requirement.txt
delta-spark        == 3.3.0
importlib_metadata == 8.7.0
load-dotenv        == 0.1.0
optional           == 0.0.1
pip                == 25.2
psutil             == 7.0.0
py4j               == 0.10.9.7
pyspark            == 3.5.6
python-dotenv      == 1.1.1
zipp               == 3.23.0